made 2nd func return long int and take long int params

+++ recursion/5-sqrt_recursion.c
@@ -6,7 +6,7 @@
  * @n: input number to get the natural square root of
  * Return: natural square root of @n
  */
-int _sqrt_recursion(int n)
+int _sqrt_recursion(long int n)
 {
 	if (n == 1)
 		return (1);
@@ -20,7 +20,7 @@ int _sqrt_recursion(int n)
  * Return: the natural square root of a number
  * If it is a decimal, the number will not be found, so it returns -1
  */
-int _sqrt_part_two(int n, int m)
+long int _sqrt_part_two(long int n, long int m)
 {
 	if (m * m == n)
 		return (m);
